MB-67520: Variety of narrowing fixes
More changes to removing narrowing warnings.

An imperfect cherry-pick of fixes mostly around collection stat code
and flatbuffer generation.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230716
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/backup/backup.cc
@@ -35,7 +35,9 @@ reEncodeOpenCollections(flatbuffers::FlatBufferBuilder& builder,
                 ScopeIDType(open.metaData.sid),
                 CollectionIDType(open.metaData.cid),
                 open.metaData.maxTtl.has_value(),
-                open.metaData.maxTtl.value_or(std::chrono::seconds(0)).count(),
+                gsl::narrow_cast<uint32_t>(
+                        open.metaData.maxTtl.value_or(std::chrono::seconds(0))
+                                .count()),
                 builder.CreateString(open.metaData.name),
                 getHistoryFromCanDeduplicate(open.metaData.canDeduplicate),
                 getMeteredFromEnum(open.metaData.metered),
